Final create of func to load ingredients cocktails and favourites cocktails
