+++ Mixly/Services/CocktailService.swift
@@ -10,4 +10,7 @@ import Foundation
 protocol CocktailService {
     func loadFilteredCocktailsBy(ids: [Int]) -> [Cocktail]
     func loadCocktailsBy(string: String) -> [Cocktail]
+    func loadFavouritesCocktails() -> [Cocktail]
+    func toggleCocktailFavourite(cocktail: Cocktail)
+    func loadIngredientsBy(string: String) -> [Ingredient]
 }
